backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
PyJWT
python-multipart